+++ open-mind/src/App.js
@@ -1,9 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Homepage from "./pages/Homepage/Homepage";
-import ButtonTest from "./components/commons/Buttons/ButtonTest";
+import Homepage from "./pages/Home/Homepage";
 import IndividualFeed from "./pages/IndividualFeed/IndividualFeed";
-import QuestionListPage from "./pages/QuestionListPage/QuestionListPage";
-import Dropdowntest from "./components/commons/Dropdown/Dropdowntest";
+import QuestionListPage from "./pages/QuestionList/QuestionListPage";
 
 function App() {
   return (
@@ -15,9 +13,6 @@ function App() {
           <Route path="post/:subjectId" element={<IndividualFeed />} />
           <Route path="post/:subjectId/answer" element={<IndividualFeed />} />
         </Route>
-        {/* test용 페이지라서 따로 빼뒀습니다 */}
-        <Route path="/buttontest" element={<ButtonTest />} />
-        <Route path="/dropdowntest" element={<Dropdowntest />} />
       </Routes>
     </BrowserRouter>
   );
