--- open-mind/src/components/Commons/Dropdown/dropdowntest.module.css
@@ -1,7 +1,0 @@
-.edit:hover {
-  fill: var(--blue50-color);
-}
-
-.close:hover {
-  fill: var(--blue50-color);
-}
