+++ open-mind/src/pages/Home/Homepage.jsx
@@ -1,21 +1,21 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import styles from "./homepage.module.css";
-import InputField from "../../components/commons/InputField/InputField";
+import InputField from "../../components/commons/Input/InputField";
 import DefaultButton from "../../components/commons/Buttons/DefaultButton";
 import { Link, useNavigate } from "react-router-dom";
-import { handleCreateFeed } from "../../api/hompage/homepage";
+import { handleCreateFeed } from "../../api/subject/subject";
 
-import logo1 from "../../assets/logo/logo1.png"
-import logo2 from "../../assets/logo/logo2.png"
-import logo3 from "../../assets/logo/logo3.png"
-import logo4 from "../../assets/logo/logo4.png"
-import logo5 from "../../assets/logo/logo5.png"
-import logo6 from "../../assets/logo/logo6.png"
-import logo7 from "../../assets/logo/logo7.png"
-import logo8 from "../../assets/logo/logo8.png"
+import logo1 from "../../assets/logo/logo1.png";
+import logo2 from "../../assets/logo/logo2.png";
+import logo3 from "../../assets/logo/logo3.png";
+import logo4 from "../../assets/logo/logo4.png";
+import logo5 from "../../assets/logo/logo5.png";
+import logo6 from "../../assets/logo/logo6.png";
+import logo7 from "../../assets/logo/logo7.png";
+import logo8 from "../../assets/logo/logo8.png";
 
-import sublogo1 from "../../assets/logo/sublogo1.png"
-import sublogo2 from "../../assets/logo/sublogo2.png"
+import sublogo1 from "../../assets/logo/sublogo1.png";
+import sublogo2 from "../../assets/logo/sublogo2.png";
 
 const go_question_button = "질문하러 가기";
 const take_question_button = "질문 받기";
