+++ open-mind/src/pages/IndividualFeed/IndividualFeed.jsx
@@ -7,9 +7,8 @@ import Toast from "../../components/commons/Toast/Toast";
 import { useState, useEffect } from "react";
 import { Helmet } from "react-helmet-async";
 import { Link, useParams, useLocation, useNavigate } from "react-router-dom";
-import { getSubjects, deleteSubjects } from "../../api/subjectApi/subjectApi";
-import { getQuestionsList } from "../../api/questionApi/questionApi";
-import { postQuestions } from "../../api/questionApi/questionApi";
+import { getSubjects, deleteSubjects } from "../../api/subject/subject";
+import { getQuestionsList, postQuestions } from "../../api/question/question";
 import FeedCard from "../../components/commons/FeedCard/FeedCard";
 import FeedCardEmpty from "../../components/commons/FeedCard/FeedCardEmpty";
 import Modal from "../../components/commons/Modal/Modal";
