+++ open-mind/src/pages/QuestionList/QuestionListPage.jsx
@@ -8,7 +8,7 @@ import ArrowDownGrayIcon from "../../assets/icon/arrow-down-gray.svg";
 import ArrowDownDarkIcon from "../../assets/icon/arrow-down.svg";
 import ArrowUpDarkIcon from "../../assets/icon/arrow-up.svg";
 import DefaultButton from "../../components/commons/Buttons/DefaultButton";
-import InputField from "../../components/commons/InputField/InputField";
+import InputField from "../../components/commons/Input/InputField";
 
 function QuestionListPage() {
   const [sortType, setSortType] = useState("최신순");
